requirements.txt
requests>=2.31.0
python-whois>=0.8.0
python-Levenshtein>=0.21.1
tldextract>=5.0.0
Pillow>=10.0.0
pycryptodome>=3.19.0
reportlab>=4.0.0
scikit-learn>=1.3.0

